refactor(app): split prompt template settings
